requirements.txt
groq
python-dotenv
beautifulsoup4
ddgs
lxml
openpyxl
pandas
plotly
requests
selenium
streamlit
webdriver-manager
# New sources
kaggle
huggingface-hub
# Robustness
httpx
tenacity
# Caching
diskcache
